authentification completed. Need to work on plaid api

+++ frontend/src/components/ExpenseDetails/ExpenseDetails.jsx
@@ -1,4 +1,5 @@
 import { useExpensesContext } from "../../hooks/useExpensesContext";
+import { useAuthContext } from "../../hooks/useAuthContext";
 
 import "./expenseDetails.css";
 
@@ -7,28 +8,35 @@ import formatDistanceToNow from "date-fns/formatDistanceToNow";
 
 const ExpenseDetails = ({ expense }) => {
   const { dispatch } = useExpensesContext();
-
+  const { user } = useAuthContext();
   const handleClick = async () => {
-    const response = await fetch("/expenses/" + expense._id, {
+    if (!user) {
+      return;
+    }
+
+    const response = await fetch("/api/expenses/" + expense._id, {
       method: "DELETE",
+      headers: {
+        Authorization: `Bearer ${user.token}`,
+      },
     });
     const json = await response.json();
 
     if (response.ok) {
-      dispatch({ type: "DELETE_EXPENSE", payload: json });
+      dispatch({ type: "DELETE_EXPENSE", PAYLOAD: json });
     }
   };
 
   return (
     <div className="expense-details">
-      <h4>{expense.title}</h4>
+      <h4>{expense.transactionName}</h4>
       <p>
-        <strong>Load (kg): </strong>
-        {expense.load}
+        <strong>Amount: </strong>
+        {expense.transactionAmount}
       </p>
       <p>
-        <strong>Reps: </strong>
-        {expense.reps}
+        <strong>Date: </strong>
+        {expense.transactionDate}
       </p>
       <p>
         {formatDistanceToNow(new Date(expense.createdAt), { addSuffix: true })}
